Add IP discovery.
Before a client wants to join a network, it may send a HTTP "ping"
request to a known participant. The participant will send a HTTP "pong"
request in response, which informs the client of its IP address.

The IP address of a client is automatically retrieved from the first
HTTP request received by Jetty.